Change type of `CFDE_TextOut::m_iCurLine` to size_t.
This member variable is used to index into an array, so it is never
negative.

Bug: pdfium:774
Reviewed-on: https://pdfium-review.googlesource.com/c/pdfium/+/95170
Commit-Queue: Lei Zhang <thestig@chromium.org>

+++ xfa/fde/cfde_textout.h
@@ -115,7 +115,7 @@ class CFDE_TextOut {
   WideString m_wsText;
   CFX_Matrix m_Matrix;
   std::deque<Line> m_ttoLines;
-  int32_t m_iCurLine = 0;
+  size_t m_iCurLine = 0;
   size_t m_iCurPiece = 0;
   int32_t m_iTotalLines = 0;
   std::vector<TextCharPos> m_CharPos;
